fixes to codebase and added example integration test

+++ src/main/java/com/learning/courses/dto/CreateCourseDTO.java
@@ -26,6 +26,6 @@ public class CreateCourseDTO implements Serializable {
     private Integer courseLength;
 
     @NotBlank
-    private Integer tutorId;
+    private Long tutorId;
 
 }
